+++ rust/stackable-cockpit/src/platform/stack/params.rs
@@ -21,3 +21,17 @@ pub struct StackInstallParameters {
     pub chart_source: ChartSourceType,
     pub operator_values: Mapping,
 }
+
+pub struct StackUninstallParameters {
+    pub stack_name: String,
+
+    /// Optional name of the demo, which is only present in case this stack is uninstalled as part of
+    /// a demo. This is unset in case a stack is uninstalled directly.
+    pub demo_name: Option<String>,
+
+    pub operator_namespace: String,
+    pub stack_namespace: String,
+
+    pub skip_operators: bool,
+    pub skip_crds: bool,
+}