+++ rust/stackable-cockpit/src/utils/k8s/client.rs
@@ -1,20 +1,23 @@
-use std::{collections::BTreeMap, string::FromUtf8Error};
+use std::{collections::BTreeMap, ops::Deref, string::FromUtf8Error};
 
 use serde::Deserialize;
 use snafu::{OptionExt, ResultExt, Snafu};
 use stackable_operator::{
     crd::listener::v1alpha1::Listener,
-    k8s_openapi::api::{
-        apps::v1::{Deployment, StatefulSet},
-        core::v1::{Endpoints, Namespace, Node, Secret, Service},
+    k8s_openapi::{
+        api::{
+            apps::v1::{Deployment, StatefulSet},
+            core::v1::{Endpoints, Namespace, Node, Secret, Service},
+        },
+        apiextensions_apiserver::pkg::apis::apiextensions::v1::CustomResourceDefinition,
     },
     kube::{
         self, Api, Discovery, ResourceExt,
-        api::{ListParams, Patch, PatchParams, PostParams},
+        api::{DeleteParams, ListParams, Patch, PatchParams, PostParams},
         core::{DynamicObject, GroupVersionKind, ObjectList, ObjectMeta, TypeMeta},
-        discovery::{ApiCapabilities, ApiResource, Scope},
+        discovery::{self, ApiCapabilities, ApiResource, Scope},
     },
-    kvp::Labels,
+    kvp::{Label, Labels},
 };
 use tokio::sync::RwLock;
 use tracing::{Span, info, instrument};
@@ -45,6 +48,11 @@ pub enum Error {
     KubeClientReplace {
         source: kube::error::Error,
         gvk: GroupVersionKind,
+    },
+
+    #[snafu(display("failed to delete object"))]
+    KubeRuntimeDelete {
+        source: kube::runtime::wait::delete::Error,
     },
 
     #[snafu(display("failed to deserialize YAML data"))]
@@ -237,6 +245,31 @@ impl Client {
         Ok(())
     }
 
+    /// Deletes CRDs for the given group suffix
+    pub async fn delete_crds_with_group_suffix(&self, group_suffix: &str) -> Result<()> {
+        let api_client = Api::<CustomResourceDefinition>::all(self.client.clone());
+
+        for crd in api_client
+            .list(&ListParams::default())
+            .await
+            .context(KubeClientFetchSnafu)?
+        {
+            if crd.spec.group.ends_with(group_suffix) {
+                if let Some(name) = crd.metadata.name {
+                    kube::runtime::wait::delete::delete_and_finalize(
+                        api_client.clone(),
+                        &name,
+                        &DeleteParams::default(),
+                    )
+                    .await
+                    .context(KubeRuntimeDeleteSnafu)?;
+                }
+            }
+        }
+
+        Ok(())
+    }
+
     /// Lists objects by looking up a GVK via the discovery. It returns an
     /// optional list of dynamic objects. The method returns `Ok(None)`
     /// if the client was unable to resolve the GVK. An error is returned
@@ -285,6 +318,86 @@ impl Client {
         Ok(Some(
             api.get(object_name).await.context(KubeClientFetchSnafu)?,
         ))
+    }
+
+    /// Deletes all objects with a given label in the provided namespace.
+    /// If no namespace is provided, deletes all clusterwide objects with the given label.
+    pub async fn delete_all_objects_with_label(
+        &self,
+        label: Label,
+        namespace: Option<&str>,
+    ) -> Result<(), Error> {
+        let api_resources = self
+            .get_api_resources()
+            .await
+            .into_iter()
+            .filter(|(_, capability)| {
+                let listing_supported = capability.supports_operation(discovery::verbs::LIST);
+
+                let capability_scope_matches = match namespace {
+                    Some(_) => capability.scope == kube::discovery::Scope::Namespaced,
+                    None => capability.scope == kube::discovery::Scope::Cluster,
+                };
+
+                listing_supported && capability_scope_matches
+            });
+
+        for (api_resource, _) in api_resources {
+            let objects = self
+                .list_objects(
+                    &GroupVersionKind {
+                        group: api_resource.group.clone(),
+                        version: api_resource.version.clone(),
+                        kind: api_resource.kind.clone(),
+                    },
+                    namespace,
+                )
+                .await?;
+
+            let Some(object_list) = objects else {
+                continue;
+            };
+
+            for object in object_list {
+                if let Some(value) = object.labels().get(&label.key().to_string()) {
+                    if value.eq(label.value().deref()) {
+                        self.delete_object(
+                            &object.metadata.name.unwrap(),
+                            &api_resource,
+                            object.metadata.namespace.as_deref(),
+                        )
+                        .await?;
+                    }
+                }
+            }
+        }
+
+        Ok(())
+    }
+
+    #[instrument(skip_all, fields(indicatif.pb_show = true))]
+    pub async fn delete_object(
+        &self,
+        object_name: &str,
+        api_resource: &ApiResource,
+        namespace: Option<&str>,
+    ) -> Result<(), Error> {
+        Span::current().pb_set_message(&format!("Deleting {} {}", api_resource.kind, object_name));
+
+        let object_api = match namespace {
+            Some(namespace) => {
+                Api::<DynamicObject>::namespaced_with(self.client.clone(), namespace, api_resource)
+            }
+            None => Api::<DynamicObject>::all_with(self.client.clone(), api_resource),
+        };
+
+        kube::runtime::wait::delete::delete_and_finalize(
+            object_api,
+            object_name,
+            &DeleteParams::foreground(),
+        )
+        .await
+        .context(KubeRuntimeDeleteSnafu)
     }
 
     /// Lists [`Service`]s by matching labels. The Services can be matched by
@@ -459,6 +572,21 @@ impl Client {
         Ok(())
     }
 
+    /// Deletes a [`Namespace`] with `name` in the cluster.
+    pub async fn delete_namespace(&self, name: String) -> Result<()> {
+        let namespace_api: Api<Namespace> = Api::all(self.client.clone());
+
+        kube::runtime::wait::delete::delete_and_finalize(
+            namespace_api,
+            &name,
+            &DeleteParams::foreground(),
+        )
+        .await
+        .context(KubeRuntimeDeleteSnafu)?;
+
+        Ok(())
+    }
+
     /// Creates a [`Namespace`] only if not already present in the current cluster.
     pub async fn create_namespace_if_needed(&self, name: String) -> Result<()> {
         if self.get_namespace(&name).await?.is_none() {
@@ -480,6 +608,16 @@ impl Client {
     pub async fn get_endpoints(&self, namespace: &str, name: &str) -> Result<Endpoints> {
         let endpoints_api: Api<Endpoints> = Api::namespaced(self.client.clone(), namespace);
         endpoints_api.get(name).await.context(KubeClientFetchSnafu)
+    }
+
+    pub async fn get_api_resources(&self) -> Vec<(ApiResource, ApiCapabilities)> {
+        let mut api_resources = Vec::new();
+
+        for group in self.discovery.read().await.groups() {
+            api_resources.append(group.recommended_resources().as_mut());
+        }
+
+        api_resources
     }
 
     /// Try to resolve the given [`GroupVersionKind`]. In case the resolution fails a discovery is run to pull in new